Revert "Add a dummy get orders function and feed it into the orders table with the right headers"
This (partially) reverts commit 8f511225e04d7a5aff3cb2faa72342b0e0805a48.

--
We keep the dummy orders fn so we can see the UI. but we remove the changes to the columns because the columns in the figma dont fully add up. So we ll review it first, then change the columns